Move decoder-related tests to decoder.rs
